Move bcdofs back to preconditioners/patch.py
Revert the relocation of bcdofs from patch.py into bcs.py; it was only
moved to support the BC handling in the objective code, which is being
dropped. Restores the original definition and imports across patch.py,
fdm.py, and bddc.py.
