+++ firedrake/bcs.py
@@ -8,7 +8,7 @@ from mpi4py import MPI
 
 import ufl
 from ufl import as_ufl, as_tensor
-from finat.ufl import MixedElement, TensorElement, VectorElement
+from finat.ufl import VectorElement
 import finat
 
 import pyop2 as op2
@@ -746,44 +746,3 @@ def restricted_function_space(V, ids):
         return spaces[0]
     else:
         return firedrake.MixedFunctionSpace(spaces, name=V.name)
-
-
-def bcdofs(bc, ghost=True):
-    # Return the global dofs fixed by a DirichletBC
-    # in the numbering given by concatenation of all the
-    # subspaces of a mixed function space
-    Z = bc.function_space()
-    while Z.parent is not None:
-        Z = Z.parent
-
-    indices = bc._indices
-    offset = 0
-
-    for (i, idx) in enumerate(indices):
-        if isinstance(Z.ufl_element(), VectorElement):
-            offset += idx
-            assert i == len(indices)-1  # assert we're at the end of the chain
-            assert Z.sub(idx).block_size == 1
-        elif isinstance(Z.ufl_element(), MixedElement):
-            if ghost:
-                offset += sum(Z.sub(j).dof_count for j in range(idx))
-            else:
-                offset += sum(Z.sub(j).dof_dset.size * Z.sub(j).block_size for j in range(idx))
-        else:
-            raise NotImplementedError("How are you taking a .sub?")
-
-        Z = Z.sub(idx)
-
-    if Z.parent is not None and isinstance(Z.parent.ufl_element(), (TensorElement, VectorElement)):
-        bs = Z.parent.block_size
-        start = 0
-        stop = 1
-    else:
-        bs = Z.block_size
-        start = 0
-        stop = bs
-    nodes = bc.nodes
-    if not ghost:
-        nodes = nodes[nodes < Z.dof_dset.size]
-
-    return np.concatenate([nodes*bs + j for j in range(start, stop)]) + offset
